apps/backend/requirements.txt
flask
flask-cors
supabase
google-generativeai
google-genai
requests
python-dotenv
